feature impletment features biz code

+++ internal/biz/features.go
@@ -2,6 +2,7 @@ package biz
 
 import (
 	"context"
+	"fmt"
 )
 
 type FeaturesRepo interface {
@@ -22,28 +23,53 @@ func NewFeaturesUsecase(repo FeaturesRepo) *FeaturesUsecase {
 	}
 }
 
+func (s *FeaturesUsecase) validateFeatures(isNew bool, features []Feature) error {
+	for _, f := range features {
+		if err := f.Validate(isNew); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 // CreateFeatures is
 func (s *FeaturesUsecase) CreateFeatures(ctx context.Context, features []Feature) error {
+	if err := s.validateFeatures(true, features); err != nil {
+		return err
+	}
 	return s.repo.CreateFeatures(ctx, features)
 }
 
 // UpdateFeatures is
 func (s *FeaturesUsecase) UpdateFeatures(ctx context.Context, features []Feature) error {
+	if err := s.validateFeatures(false, features); err != nil {
+		return err
+	}
 	return s.repo.UpdateFeatures(ctx, features)
 }
 
 // DeleteFeatures is
 func (s *FeaturesUsecase) DeleteFeatures(ctx context.Context, ids []string) error {
+	if len(ids) == 0 {
+		return fmt.Errorf("EmptyIds")
+	}
 	return s.repo.DeleteFeatures(ctx, ids)
 }
 
 // GetFeatures is
 func (s *FeaturesUsecase) GetFeatures(ctx context.Context, id string) (Feature, error) {
+	if len(id) == 0 {
+		return Feature{}, fmt.Errorf("EmptyId")
+	}
 	return s.repo.GetFeatures(ctx, id)
 }
 
 // ListFeatures is
 func (s *FeaturesUsecase) ListFeatures(ctx context.Context,
 	filter *ListFeaturesFilter) ([]Feature, error) {
+
+	if err := filter.Validate(); err != nil {
+		return nil, err
+	}
 	return s.repo.ListFeatures(ctx, filter)
 }
